refactor(coordinator): extract ASR engine wiring into coordinator/asr_wiring.rs — behavior unchanged, 446 tests green
Pure move of 23 free functions + 2 enums (QaAsrStart, AsrReleaseSession) covering
credential/permission gates, local-ASR release scheduling, provider helpers, and
build_qa_asr_start, from coordinator.rs into a new sibling submodule. Functions and
QaAsrStart impl methods made pub(super); reached from the parent via
'use asr_wiring::*;'. No logic/signature/public-API changes; coordinator.rs
5807→5285 lines.

cargo check: 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
